feat: sheet miktarlari ondalikli olabilsin, sayim sablonuna SKT/Ek Ozellik eklendi
Ham Giris ve Ham Sayim formlarinda Sheet miktari artik 1,5 gibi ondalikli
girilebiliyor (strip miktarlari yine tam sayi olarak yuvarlanarak hesaplaniyor).
raw_stock_lots.sheet_giren ve raw_stock_count_items.sayilan_sheet DECIMAL(10,2)
oldu (migration 018).

Sayim Sablonu Excel exportuna SKT ve Ek Ozellik sutunlari eklendi.

Ham Giris formunda Sheet alanindan Tab ile SKT alanina gecerken odak kayboluyordu
(satirin tamami yeniden render edildigi icin) - artik render sonrasi SKT alanina
odak geri veriliyor. SKT alaninda Tab da art0 Enter gibi yeni kalem satiri ekliyor.

+++ api/stok/ham-lotlar/index.php
@@ -16,7 +16,7 @@ function lotResponse(array $row): array {
         'cutoff'      => $row['cutoff'],
         'ekOzellik'   => $row['ek_ozellik'],
         'kategoriId'  => $row['kategori_id'],
-        'sheetGiren'  => (int)$row['sheet_giren'],
+        'sheetGiren'  => (float)$row['sheet_giren'],
         'stripGiren'  => (int)$row['strip_giren'],
         'mevcutStrip' => (int)$row['mevcut_strip'],
         'sktTarih'    => $row['skt_tarih'],
